feat(login): Added login functionality back pt. 1
Custom Games + button is not aligned correctly
Refactoring min_window, max_window and close_window
::selector is broken
